requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-scheduler==0.10.1
et-xmlfile==1.1.0
frozenlist==1.4.1
icalendar==5.0.11
idna==3.6
inexactsearch==1.0.2
multidict==6.0.4
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.1.0
pyspellchecker==0.8.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qq-botpy==1.2.0
qqbot==2.3.11
requests==2.31.0
silpa-common==0.3
six==1.16.0
soundex==1.1.3
spellchecker==0.4
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4

gino~=1.0.1